+++ src/apps/shared/interfaces/common.ts
@@ -1,9 +1,0 @@
-export interface FocusedApp {
-  hwnd: number;
-  name: string;
-  title: string;
-  exe: string | null;
-  umid: string | null;
-  isMaximized: boolean;
-  isSeelenOverlay: boolean;
-}