+++ src/background/seelen_weg/hook.rs
@@ -40,13 +40,15 @@ impl SeelenWeg {
                     Self::update_app(window)?;
                 }
             }
-            /* WinEvent::ObjectNameChange => {
+            WinEvent::ObjectNameChange
+            | WinEvent::SystemMinimizeStart
+            | WinEvent::SystemMinimizeEnd
+            | WinEvent::SyntheticMaximizeStart
+            | WinEvent::SyntheticMaximizeEnd => {
                 if Self::contains_app(window) {
                     Self::update_app(window)?;
-                } else if Self::should_be_added(window) {
-                    Self::add(window)?;
                 }
-            } */
+            }
             _ => {}
         }
         Ok(())